Proposed restructure
