+++ gym_minigrid/envs/unlockpickup.py
@@ -1,5 +1,5 @@
+from gym_minigrid.core.roomgrid import RoomGrid
 from gym_minigrid.minigrid import COLOR_NAMES, MissionSpace
-from gym_minigrid.roomgrid import RoomGrid
 
 
 class UnlockPickupEnv(RoomGrid):
