+++ gym_minigrid/utils/rendering.py
