+++ gym_minigrid/utils/window.py
